PLATUI-301 Json reads for int-like strings
For JSON objects that have fields that are
integers that are typed as strings, e.g.
"1", we should parse as an 'IntString'
rather than failing to read this as a
JsNumber. Equally, numbers such as 1
can also be parsed as an 'IntString'.

Hence, both an appropriate integer
JsNumber or JsString can be parsed;
else the field will fail to be
parsed with an appropriate JsError.

Parsing as an 'IntString' allows
the error to be attached to the
relevant field inside the JsError
rather than an error related to
mapping inappropriate strings to
ints after deserialisation, which
gives a more confusing stack trace.

This functionality allows us to
parse JsObject fields for components
in the same way GovUK frontend Nunjucks
does, not minding about whether an
integer-like field is a string or
integer initially.