PLATUI-904: Use parameters for customising timeout dialog content
We explored wiring up messages in the library but decided to stick with the existing parameter-based API
for simplicity and consistency until such time as we would be defaulting content in the library itself. As it stands
the content defaults are stored in the Javascript in hmrc/hmrc-frontend and most services
are using the defaults. Adding in wiring for messages would mean adding checks for message key
existence (via messages.isDefinedAt) to avoid silent defaulting to the message key itself and we would need to
settle on a message key prefix unlikely to be used by any existing services.